refactor: MeasurPolicy 재사용 가능하도록 변경
